chore(maintenance): move to /packages
