+++ packages/administration/src/maintenance/index.ts
@@ -1,0 +1,2 @@
+export { MaintenanceModule } from './maintenance.module';
+export { MaintenanceComponent } from './maintenance.component';
