--- src/app/admin/maintenance/index.ts
@@ -1,1 +1,0 @@
-export { MaintenanceModule } from './maintenance.module';
